Add data.csv to webserver
Add handle to download the stored data as a csv file. The index page will
use this draw a graph.

The data is served using chunked response, because building the entire
csv file uses too much RAM. Several lines are put together in a chunk,
to reduce the overhead of many calls to sendContent().

The time column is time for the sample relative to now in seconds, this
is quick to calculate, short to send and can be intepreted in javascript
by the index page.